requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.9.0
matplotlib>=3.5.0
tensorboard>=2.10.0
pandas>=1.4.0
tqdm>=4.64.0
pyyaml>=6.0
h5py>=3.7.0
ray[serve]>=2.0.0
optuna>=3.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
scikit-learn>=1.1.0

